requirements-docker.txt
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
sse-starlette>=1.6.0
pypdf>=3.0.0
python-docx>=0.8.0
openpyxl>=3.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
